spanner/demo/requirements.txt

google-cloud-spanner==3.44.0
google-cloud-storage==2.16.0
pandas==2.2.1